r/aws_glue_dev_endpoint: Reduce visibility.
