+++ internal/service/glue/service_package_gen.go
@@ -109,7 +109,7 @@ func (p *servicePackage) SDKResources(ctx context.Context) []*types.ServicePacka
 			},
 		},
 		{
-			Factory:  ResourceDevEndpoint,
+			Factory:  resourceDevEndpoint,
 			TypeName: "aws_glue_dev_endpoint",
 			Name:     "Dev Endpoint",
 			Tags: &types.ServicePackageResourceTags{
